deployment + scripts

+++ typechain-types/factories/src/interfaces/ICondition__factory.ts
@@ -17,14 +17,14 @@ const _abi = [
         type: "address",
       },
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "id",
-        type: "uint16",
+        type: "uint32",
       },
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "action",
-        type: "uint16",
+        type: "uint32",
       },
     ],
     name: "actionValid",
@@ -41,14 +41,14 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "id",
-        type: "uint16",
+        type: "uint32",
       },
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "action",
-        type: "uint16",
+        type: "uint32",
       },
     ],
     name: "addAutomationToCondition",
@@ -65,14 +65,14 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "id",
-        type: "uint16",
+        type: "uint32",
       },
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "action",
-        type: "uint16",
+        type: "uint32",
       },
     ],
     name: "addStrategyToCondition",
@@ -94,9 +94,9 @@ const _abi = [
         type: "address",
       },
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "id",
-        type: "uint16",
+        type: "uint32",
       },
     ],
     name: "checkCondition",
@@ -118,9 +118,9 @@ const _abi = [
         type: "address",
       },
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "id",
-        type: "uint16",
+        type: "uint32",
       },
     ],
     name: "conditionActive",
@@ -137,9 +137,9 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "id",
-        type: "uint16",
+        type: "uint32",
       },
     ],
     name: "deleteCondition",
@@ -155,9 +155,9 @@ const _abi = [
         type: "address",
       },
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "id",
-        type: "uint16",
+        type: "uint32",
       },
     ],
     name: "isUpdateable",
@@ -174,14 +174,14 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "id",
-        type: "uint16",
+        type: "uint32",
       },
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "automation",
-        type: "uint16",
+        type: "uint32",
       },
     ],
     name: "removeAutomationFromCondition",
@@ -198,14 +198,14 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "id",
-        type: "uint16",
+        type: "uint32",
       },
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "strategy",
-        type: "uint16",
+        type: "uint32",
       },
     ],
     name: "removeStrategyFromCondition",
@@ -227,14 +227,14 @@ const _abi = [
         type: "address",
       },
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "id",
-        type: "uint16",
+        type: "uint32",
       },
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "strategy",
-        type: "uint16",
+        type: "uint32",
       },
     ],
     name: "strategyValid",
@@ -251,9 +251,9 @@ const _abi = [
   {
     inputs: [
       {
-        internalType: "uint16",
+        internalType: "uint32",
         name: "id",
-        type: "uint16",
+        type: "uint32",
       },
     ],
     name: "updateCondition",
